1.0.0-RC1 Effort Started
This new major version aims to bring a new and hopefully cleaner project structure.
The `scalaz` based configuration is moved to a different module to isolate better from the core.
In the future a `cats` based version will be added as well.
More core utilities were added and the old ones were brushed up for better consistency and clarity.